refactor(aws_service): move threading call func
