+++ prowler/providers/aws/services/accessanalyzer/accessanalyzer_service.py
@@ -1,4 +1,3 @@
-import threading
 from typing import Optional
 
 from botocore.exceptions import ClientError
@@ -18,18 +17,6 @@ class AccessAnalyzer(AWS_Service):
         self.__threading_call__(self.__list_analyzers__)
         self.__list_findings__()
         self.__get_finding_status__()
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __list_analyzers__(self, regional_client):
         logger.info("AccessAnalyzer - Listing Analyzers...")
