+++ prowler/providers/aws/services/account/account_service.py
@@ -7,8 +7,5 @@ class Account(AWS_Service):
         # Call AWS_Service's __init__
         super().__init__(__class__.__name__, audit_info)
 
-    def __get_session__(self):
-        return self.session
-
 
 ### This service don't need boto3 calls
