+++ prowler/providers/aws/services/acm/acm_service.py
@@ -1,4 +1,3 @@
-import threading
 from datetime import datetime
 from typing import Optional
 
@@ -18,18 +17,6 @@ class ACM(AWS_Service):
         self.__threading_call__(self.__list_certificates__)
         self.__describe_certificates__()
         self.__list_tags_for_certificate__()
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __list_certificates__(self, regional_client):
         logger.info("ACM - Listing Certificates...")
