+++ prowler/providers/aws/services/autoscaling/autoscaling_service.py
@@ -1,5 +1,3 @@
-import threading
-
 from pydantic import BaseModel
 
 from prowler.lib.logger import logger
@@ -16,18 +14,6 @@ class AutoScaling(AWS_Service):
         self.__threading_call__(self.__describe_launch_configurations__)
         self.groups = []
         self.__threading_call__(self.__describe_auto_scaling_groups__)
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __describe_launch_configurations__(self, regional_client):
         logger.info("AutoScaling - Describing Launch Configurations...")
