+++ prowler/providers/aws/services/awslambda/awslambda_service.py
@@ -1,6 +1,5 @@
 import io
 import json
-import threading
 import zipfile
 from enum import Enum
 from typing import Any, Optional
@@ -33,18 +32,6 @@ class Lambda(AWS_Service):
 
         self.__threading_call__(self.__get_policy__)
         self.__threading_call__(self.__get_function_url_config__)
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __list_functions__(self, regional_client):
         logger.info("Lambda - Listing Functions...")
