+++ prowler/providers/aws/services/cloudformation/cloudformation_service.py
@@ -1,4 +1,3 @@
-import threading
 from typing import Optional
 
 from botocore.client import ClientError
@@ -17,18 +16,6 @@ class CloudFormation(AWS_Service):
         self.stacks = []
         self.__threading_call__(self.__describe_stacks__)
         self.__describe_stack__()
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __describe_stacks__(self, regional_client):
         """Get ALL CloudFormation Stacks"""
