+++ prowler/providers/aws/services/cloudtrail/cloudtrail_service.py
@@ -1,4 +1,3 @@
-import threading
 from datetime import datetime
 from typing import Optional
 
@@ -21,18 +20,6 @@ class Cloudtrail(AWS_Service):
         self.__get_insight_selectors__()
         self.__get_event_selectors__()
         self.__list_tags_for_resource__()
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __get_trails__(self, regional_client):
         logger.info("Cloudtrail - Getting trails...")
