+++ prowler/providers/aws/services/codebuild/codebuild_service.py
@@ -1,5 +1,4 @@
 import datetime
-import threading
 from dataclasses import dataclass
 from typing import Optional
 
@@ -16,18 +15,6 @@ class Codebuild(AWS_Service):
         self.projects = []
         self.__threading_call__(self.__list_projects__)
         self.__list_builds_for_project__()
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __list_projects__(self, regional_client):
         logger.info("Codebuild - listing projects")
