+++ prowler/providers/aws/services/directoryservice/directoryservice_service.py
@@ -1,4 +1,3 @@
-import threading
 from datetime import datetime
 from enum import Enum
 from typing import Optional, Union
@@ -23,18 +22,6 @@ class DirectoryService(AWS_Service):
         self.__threading_call__(self.__list_certificates__)
         self.__threading_call__(self.__get_snapshot_limits__)
         self.__list_tags_for_resource__()
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __describe_directories__(self, regional_client):
         logger.info("DirectoryService - Describing Directories...")
