+++ prowler/providers/aws/services/ec2/ec2_service.py
@@ -1,4 +1,3 @@
-import threading
 from datetime import datetime
 from typing import Optional
 
@@ -37,18 +36,6 @@ class EC2(AWS_Service):
         self.__threading_call__(self.__get_ebs_encryption_by_default__)
         self.elastic_ips = []
         self.__threading_call__(self.__describe_addresses__)
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __describe_instances__(self, regional_client):
         logger.info("EC2 - Describing EC2 Instances...")
