+++ prowler/providers/aws/services/ecr/ecr_service.py
@@ -1,4 +1,3 @@
-import threading
 from datetime import datetime
 from json import loads
 from typing import Optional
@@ -24,18 +23,6 @@ class ECR(AWS_Service):
         self.__threading_call__(self.__get_repository_lifecycle_policy__)
         self.__threading_call__(self.__get_registry_scanning_configuration__)
         self.__threading_call__(self.__list_tags_for_resource__)
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __describe_registries_and_repositories__(self, regional_client):
         logger.info("ECR - Describing registries and repositories...")
