+++ prowler/providers/aws/services/efs/efs_service.py
@@ -1,5 +1,4 @@
 import json
-import threading
 from typing import Optional
 
 from botocore.client import ClientError
@@ -18,18 +17,6 @@ class EFS(AWS_Service):
         self.filesystems = []
         self.__threading_call__(self.__describe_file_systems__)
         self.__describe_file_system_policies__()
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __describe_file_systems__(self, regional_client):
         logger.info("EFS - Describing file systems...")
