+++ prowler/providers/aws/services/elbv2/elbv2_service.py
@@ -1,4 +1,3 @@
-import threading
 from typing import Optional
 
 from botocore.client import ClientError
@@ -21,18 +20,6 @@ class ELBv2(AWS_Service):
         self.__threading_call__(self.__describe_load_balancer_attributes__)
         self.__threading_call__(self.__describe_rules__)
         self.__describe_tags__()
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __describe_load_balancers__(self, regional_client):
         logger.info("ELBv2 - Describing load balancers...")
