+++ prowler/providers/aws/services/emr/emr_service.py
@@ -1,4 +1,3 @@
-import threading
 from enum import Enum
 from typing import Optional
 
@@ -20,18 +19,6 @@ class EMR(AWS_Service):
         self.__threading_call__(self.__list_clusters__)
         self.__threading_call__(self.__describe_cluster__)
         self.__threading_call__(self.__get_block_public_access_configuration__)
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __list_clusters__(self, regional_client):
         logger.info("EMR - Listing Clusters...")
