+++ prowler/providers/aws/services/glacier/glacier_service.py
@@ -1,5 +1,4 @@
 import json
-import threading
 from typing import Optional
 
 from botocore.client import ClientError
@@ -19,18 +18,6 @@ class Glacier(AWS_Service):
         self.__threading_call__(self.__list_vaults__)
         self.__threading_call__(self.__get_vault_access_policy__)
         self.__list_tags_for_vault__()
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __list_vaults__(self, regional_client):
         logger.info("Glacier - Listing Vaults...")
