+++ prowler/providers/aws/services/glue/glue_service.py
@@ -1,4 +1,3 @@
-import threading
 from typing import Optional
 
 from pydantic import BaseModel
@@ -25,18 +24,6 @@ class Glue(AWS_Service):
         self.__threading_call__(self.__get_security_configurations__)
         self.jobs = []
         self.__threading_call__(self.__get_jobs__)
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __get_connections__(self, regional_client):
         logger.info("Glue - Getting connections...")
