+++ prowler/providers/aws/services/guardduty/guardduty_service.py
@@ -1,4 +1,3 @@
-import threading
 from typing import Optional
 
 from pydantic import BaseModel
@@ -20,18 +19,6 @@ class GuardDuty(AWS_Service):
         self.__list_members__()
         self.__get_administrator_account__()
         self.__list_tags_for_resource__()
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __list_detectors__(self, regional_client):
         logger.info("GuardDuty - listing detectors...")
