+++ prowler/providers/aws/services/inspector2/inspector2_service.py
@@ -1,5 +1,3 @@
-import threading
-
 from pydantic import BaseModel
 
 from prowler.lib.logger import logger
@@ -15,18 +13,6 @@ class Inspector2(AWS_Service):
         self.inspectors = []
         self.__threading_call__(self.__batch_get_account_status__)
         self.__list_findings__()
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __batch_get_account_status__(self, regional_client):
         # We use this function to check if inspector2 is enabled
