+++ prowler/providers/aws/services/kms/kms_service.py
@@ -1,5 +1,4 @@
 import json
-import threading
 from typing import Optional
 
 from pydantic import BaseModel
@@ -21,18 +20,6 @@ class KMS(AWS_Service):
             self.__get_key_rotation_status__()
             self.__get_key_policy__()
             self.__list_resource_tags__()
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __list_keys__(self, regional_client):
         logger.info("KMS - Listing Keys...")
