+++ prowler/providers/aws/services/macie/macie_service.py
@@ -1,5 +1,3 @@
-import threading
-
 from pydantic import BaseModel
 
 from prowler.lib.logger import logger
@@ -13,18 +11,6 @@ class Macie(AWS_Service):
         super().__init__("macie2", audit_info)
         self.sessions = []
         self.__threading_call__(self.__get_macie_session__)
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __get_macie_session__(self, regional_client):
         logger.info("Macie - Get Macie Session...")
