+++ prowler/providers/aws/services/rds/rds_service.py
@@ -1,4 +1,3 @@
-import threading
 from typing import Optional
 
 from botocore.client import ClientError
@@ -27,18 +26,6 @@ class RDS(AWS_Service):
         self.__threading_call__(self.__describe_db_cluster_snapshots__)
         self.__threading_call__(self.__describe_db_cluster_snapshot_attributes__)
         self.__threading_call__(self.__describe_db_engine_versions__)
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __describe_db_instances__(self, regional_client):
         logger.info("RDS - Describe Instances...")
