+++ prowler/providers/aws/services/redshift/redshift_service.py
@@ -1,4 +1,3 @@
-import threading
 from typing import Optional
 
 from pydantic import BaseModel
@@ -17,18 +16,6 @@ class Redshift(AWS_Service):
         self.__threading_call__(self.__describe_clusters__)
         self.__describe_logging_status__(self.regional_clients)
         self.__describe_cluster_snapshots__(self.regional_clients)
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __describe_clusters__(self, regional_client):
         logger.info("Redshift - describing clusters...")
