+++ prowler/providers/aws/services/secretsmanager/secretsmanager_service.py
@@ -1,4 +1,3 @@
-import threading
 from typing import Optional
 
 from pydantic import BaseModel
@@ -15,18 +14,6 @@ class SecretsManager(AWS_Service):
         super().__init__(__class__.__name__, audit_info)
         self.secrets = {}
         self.__threading_call__(self.__list_secrets__)
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __list_secrets__(self, regional_client):
         logger.info("SecretsManager - Listing Secrets...")
