+++ prowler/providers/aws/services/securityhub/securityhub_service.py
@@ -1,5 +1,3 @@
-import threading
-
 from botocore.client import ClientError
 from pydantic import BaseModel
 
@@ -15,18 +13,6 @@ class SecurityHub(AWS_Service):
         super().__init__(__class__.__name__, audit_info)
         self.securityhubs = []
         self.__threading_call__(self.__describe_hub__)
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __describe_hub__(self, regional_client):
         logger.info("SecurityHub - Describing Hub...")
