+++ prowler/providers/aws/services/sqs/sqs_service.py
@@ -1,4 +1,3 @@
-import threading
 from json import loads
 from typing import Optional
 
@@ -18,18 +17,6 @@ class SQS(AWS_Service):
         self.__threading_call__(self.__list_queues__)
         self.__get_queue_attributes__(self.regional_clients)
         self.__list_queue_tags__()
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __list_queues__(self, regional_client):
         logger.info("SQS - describing queues...")
