+++ prowler/providers/aws/services/ssm/ssm_service.py
@@ -1,5 +1,4 @@
 import json
-import threading
 from enum import Enum
 from typing import Optional
 
@@ -24,18 +23,6 @@ class SSM(AWS_Service):
         self.__threading_call__(self.__describe_document_permission__)
         self.__threading_call__(self.__list_resource_compliance_summaries__)
         self.__threading_call__(self.__describe_instance_information__)
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __list_documents__(self, regional_client):
         logger.info("SSM - Listing Documents...")
