+++ prowler/providers/aws/services/ssmincidents/ssmincidents_service.py
@@ -1,5 +1,3 @@
-import threading
-
 from botocore.client import ClientError
 from pydantic import BaseModel
 
@@ -25,18 +23,6 @@ class SSMIncidents(AWS_Service):
         self.response_plans = []
         self.__threading_call__(self.__list_response_plans__)
         self.__list_tags_for_resource__()
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __list_replication_sets__(self):
         logger.info("SSMIncidents - Listing Replication Sets...")
