+++ prowler/providers/aws/services/wafv2/wafv2_service.py
@@ -1,5 +1,3 @@
-import threading
-
 from pydantic import BaseModel
 
 from prowler.lib.logger import logger
@@ -15,18 +13,6 @@ class WAFv2(AWS_Service):
         self.web_acls = []
         self.__threading_call__(self.__list_web_acls__)
         self.__threading_call__(self.__list_resources_for_web_acl__)
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __list_web_acls__(self, regional_client):
         logger.info("WAFv2 - Listing Regional Web ACLs...")
