+++ prowler/providers/aws/services/wellarchitected/wellarchitected_service.py
@@ -1,4 +1,3 @@
-import threading
 from typing import Optional
 
 from pydantic import BaseModel
@@ -16,18 +15,6 @@ class WellArchitected(AWS_Service):
         self.workloads = []
         self.__threading_call__(self.__list_workloads__)
         self.__list_tags_for_resource__()
-
-    def __get_session__(self):
-        return self.session
-
-    def __threading_call__(self, call):
-        threads = []
-        for regional_client in self.regional_clients.values():
-            threads.append(threading.Thread(target=call, args=(regional_client,)))
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
 
     def __list_workloads__(self, regional_client):
         logger.info("WellArchitected - Listing Workloads...")
